% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_df_to_sf.R
\name{cast_df_to_sf}
\alias{cast_df_to_sf}
\title{Transform data.frame with coordinates to sf data frame}
\usage{
cast_df_to_sf(df = NULL, crs = 4326, ...)
}
\arguments{
\item{df}{(required, data.frame) Dataframe with coordinate columns. Default: NULL}

\item{crs}{(optional, integer) Coordinate reference system as a valid EPSG value. Default: 4326 (WGS84)}

\item{...}{(optional) For internal arguments only.}
}
\value{
sf data frame with point geometry
}
\description{
Transform data.frame with coordinates to sf data frame
}
\seealso{
Other casting_functions: 
\code{\link{cast_sf_to_bbox}()},
\code{\link{cast_sf_to_xy}()},
\code{\link{cast_xy_to_xyz}()}
}
\concept{casting_functions}
