% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_sf_to_xy.R
\name{cast_sf_to_xy}
\alias{cast_sf_to_xy}
\title{Transform sf data.frame into xy matrix}
\usage{
cast_sf_to_xy(df = NULL, ...)
}
\arguments{
\item{df}{(required, sf) An sf data frame with spatial geometry}

\item{...}{Internal parameters passed from parent functions.}
}
\value{
Numeric matrix with columns "x" and "y"
}
\description{
Transform sf data.frame into xy matrix
}
\seealso{
Other casting_functions: 
\code{\link{cast_df_to_sf}()},
\code{\link{cast_sf_to_bbox}()},
\code{\link{cast_xy_to_xyz}()}
}
\concept{casting_functions}
