% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_fold_plot.R
\name{spatial_fold_plot}
\alias{spatial_fold_plot}
\title{Quick plot of a single training fold}
\usage{
spatial_fold_plot(
  xy = NULL,
  center = NULL,
  training_fold = NULL,
  color_training = "blue3",
  color_testing = "red3",
  legend_position = "bottomleft"
)
}
\arguments{
\item{training_fold}{(required, logical vector) Output of \link{training_fold} on \code{xy}. Default: \code{NULL}}

\item{color_training}{(optional, color name) Color of the training records. Default: \code{"blue3"}}

\item{color_testing}{(optional, color name) Color of the testing records. Default: \code{"red3"}}

\item{legend_position}{(optional, string) Position of the legend. One of "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center". Default: \code{"bottomleft"}}
}
\value{
invisible
}
\description{
Quick plot of a single training fold
}
\examples{
training <- method_contiguous(
  xy = xy_matrix,
  center = 1,
  step_x = 0.4,
  step_y = 0.1,
  target = 15000
)

spatial_fold_plot(
  xy = xy_matrix,
  center = 1,
  training_fold = training
)
}
