% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_fold_plot.R
\name{training_fold_plot}
\alias{training_fold_plot}
\title{Quick plot of a single training fold}
\usage{
training_fold_plot(
  xy = NULL,
  center = NULL,
  training_fold = NULL,
  color_training = "blue3",
  color_testing = "red3",
  legend_position = "bottomleft"
)
}
\arguments{
\item{xy}{(required, matrix) Two columns matrix with the locations to arrange. The first column is interpreted as "x" (longitude) and the second as "y" (latitude). Default: \code{NULL}}

\item{center}{(required, integer) Index of the training fold center. Default: \code{NULL}}

\item{training_fold}{(required, logical vector) Output of \link{training_fold} on \code{xy}. Default: \code{NULL}}

\item{color_training}{(optional, color name) Color of the training records. Default: \code{"blue3"}}

\item{color_testing}{(optional, color name) Color of the testing records. Default: \code{"red3"}}

\item{legend_position}{(optional, string) Position of the legend. One of "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center". Default: \code{"bottomleft"}}
}
\value{
invisible
}
\description{
Quick plot of a single training fold
}
\examples{
training <- training_fold(
  xy = xy_matrix,
  center = 1,
  step_x = 0.4,
  step_y = 0.1,
  training_fraction = 0.50
)

training_fold_plot(
  xy = xy_matrix,
  center = 1,
  training_fold = training
)
}
