% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xy_matrix}
\alias{xy_matrix}
\title{Matrix with 30k pairs of coordiantes}
\format{
Matrix with columns "x" and "y"
}
\usage{
data(xy_matrix)
}
\description{
Matrix with 30k pairs of coordiantes
}
\keyword{datasets}
