% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xy_sf}
\alias{xy_sf}
\title{Sf dataframe with 30k points and CRS 4326}
\format{
Matrix with columns "id" and "geometry"
}
\usage{
data(xy_sf)
}
\description{
Sf dataframe with 30k points and CRS 4326
}
\keyword{datasets}
