% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialFolds-package.R
\docType{package}
\name{spatialFolds-package}
\alias{spatialFolds}
\alias{spatialFolds-package}
\title{spatialFolds: Cohesive training and testing folds for spatial cross-validation}
\description{
Generates training and testing folds for spatial cross-validation preserving the spatial structure of the data.
}
\author{
\strong{Maintainer}: First Last \email{blasbenito@gmail.com}

}
\keyword{internal}
