% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arg_distance.R
\name{validate_arg_distance}
\alias{validate_arg_distance}
\title{Validates Argument \code{distance} in \code{spatial_thinning()}}
\usage{
validate_arg_distance(df = NULL, distance = NULL, function_name = NULL)
}
\arguments{
\item{df}{(required, sf or data.frame) Spatial data with point geometries.
If data.frame, must have recognizable coordinate columns (x/lon/longitude
and y/lat/latitude). Will be converted to sf with CRS 4326. Default: NULL}

\item{distance}{(optional, numeric) Incompatible with \code{target}. Minimum distance between nearby points. Must be in same units as coordinates. Auto-calculated (see Details) when \code{NULL} and \code{target = NULL}. Default: \code{NULL} (auto-calculated)}

\item{function_name}{(optional, string) Name of the calling function}
}
\value{
numeric
}
\description{
Validates Argument \code{distance} in \code{spatial_thinning()}
}
\examples{
data(xy_sf)
x <- validate_arg_distance(
  df = xy_sf,
  distance = NULL
  )
}
\seealso{
Other arg_validation: 
\code{\link{validate_arg_blocks}()},
\code{\link{validate_arg_sf}()},
\code{\link{validate_arg_target}()},
\code{\link{validate_arg_xy}()}
}
\concept{arg_validation}
