% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arg_target.R
\name{validate_arg_target}
\alias{validate_arg_target}
\title{Validates Argument \code{target} in \code{spatial_thinning()}}
\usage{
validate_arg_target(df = NULL, target = NULL, function_name = NULL)
}
\arguments{
\item{df}{(required, sf or data.frame) Spatial data with point geometries.
If data.frame, must have recognizable coordinate columns (x/lon/longitude
and y/lat/latitude). Will be converted to sf with CRS 4326. Default: NULL}

\item{target}{(optional, integer) Incompatible with \code{distance}. Number of points to retain (exact count not guaranteed). Must be between 1 and \code{nrow(df) - 1}. Default: \code{NULL}}

\item{function_name}{(optional, string) Name of the calling function}
}
\value{
integer
}
\description{
Validates Argument \code{target} in \code{spatial_thinning()}
}
\examples{
data(xy_sf)
x <- validate_arg_target(
  df = xy_sf,
  target = 10
  )
}
\seealso{
Other arg_validation: 
\code{\link{validate_arg_blocks}()},
\code{\link{validate_arg_distance}()},
\code{\link{validate_arg_sf}()},
\code{\link{validate_arg_xy}()}
}
\concept{arg_validation}
